requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
frozendict==2.3.7
html5lib==1.1
idna==3.4
lxml==4.9.2
multitasking==0.0.11
numpy==1.24.2
pandas==2.0.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
tabulate==0.9.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
Wikipedia-API==0.5.8
yfinance==0.2.17
